fix(triton): int64 dropout RNG offsets above 2^31 score elements (#352)
* fix(triton): int64 dropout RNG offsets above 2^31 score elements

The Triton dropout mask computed its logical element offset
`((b*H + h) * Nq) * Nkv + q*Nkv + k` in int32 (program_id and offs_*
are 32-bit), so the philox counter wrapped once B*H*Nq*Nkv exceeded
2^31 and the mask diverged from SDPA/PyTorch-main semantics. Promote
the offset math to int64 in the three generation sites (shared forward
helper, split-kv decode stage1, backward dropout multiplier); values
below 2^31 are bit-identical, verified by a stashed-baseline run.

Add large-N regressions: B1 H16 N16384 (exactly 2^32 elements) checks
each backend against the still-valid SDPA reference, and B1 H32 N16384
(2^33 elements) cross-checks cuda vs triton because torch's
mem-efficient SDPA wraps its own dropout offset in 32-bit at that
scale (fixed upstream on PyTorch main via
gemm_kernel_utils::dropout_rng_offset).


* bench: attribute >2^32 dropout parity misses to the torch reference

torch's mem-efficient SDPA derives the per-(b,h) dropout Philox offset
in 32-bit arithmetic, so the reference mask itself wraps once
B*Hq*Nq*Nkv exceeds 2^32 score elements (fixed upstream on PyTorch
main by gemm_kernel_utils::dropout_rng_offset taking int64_t extents).
FFPA keeps the correct int64 layout on every backend, so a dropout
parity miss at that scale indicts the reference, not the kernels.

The runners now detect that condition, print an explicit warning, tag
the case with [torch-ref-2^32-bug], and the Markdown marker renders
"⚠️" instead of "❌" so N=16384 dropout runs stop reporting false
regressions. Verified: N=16384 dropout (fwd and bwd) warns and tags;
N=8192 stays silent; the actual O_err is still printed for inspection.


* docs(rfc): close FC-11 - dropout parity root-caused to torch ref uint32 wrap

Rewrite the FC-11 entry from Draft to Done with the verified root
cause: the bench failure was a stale .so (native source was already
correct since #199), Triton's RNG offset math genuinely wrapped in
int32 (fixed in 542f774), and the local torch 2.11.0 mem-efficient
SDPA reference wraps its own dropout offset in uint32 above 2^32
score elements (fixed upstream on PyTorch main). Record the five-way
cross-impl validation, the bench attribution tags from e1fe363, and
the large-N regression tests. Also note the deferred tma+dropout
performance anomaly. Sync the completion checklist, overview table,
and SKILL.md section 7.2 RNG semantics.


* bench: print torch-ref 2^32 dropout warn only under --verbose

The per-case [warn] about the torch mem-efficient SDPA dropout RNG offset
wrapping above 2^32 score elements now prints only when --verbose is set.
The Markdown ⚠️ marker and the inline [torch-ref-2^32-bug] note (also
verbose-only) keep tracking the affected rows.


* docs(rfc): add sub-item numbering and split PC-0 by kernel family

RFC now supports child items under any major FC/PC (parent-child ids like
PC-0-0, indented rows/checkboxes, parent status aggregates child progress).

PC-0 (attn mask performance) splits into three scene-scoped children:
- PC-0-0 cute/cute_tma (fp16 cute family, warmup step for plan A)
- PC-0-1 fp8/fp4 (quantized six-kernel family, former PC-0 body)
- PC-0-2 native/native_tma (scalar loader path, design sketch pending)

Only PC-0 gains children for now; other items unchanged.


---------

+++ src/ffpa_attn/triton/_ffpa_bwd.py
@@ -211,8 +211,11 @@ def _dropout_multiplier(
   mult = tl.full([offs_m.shape[0], offs_n.shape[0]], 1.0, dtype=tl.float32)
   if HAS_DROPOUT:
     # Replay the exact forward dropout mask using SDPA's logical score layout.
-    linear = off_hb * seqlen_q * seqlen_k + offs_m[:, None] * seqlen_k + offs_n[
-      None, :]
+    # int64: B*H*Nq*Nkv exceeds 2**31 for realistic prefill shapes.
+    linear = (
+      off_hb.to(tl.int64) * seqlen_q * seqlen_k +
+      offs_m[:, None].to(tl.int64) * seqlen_k + offs_n[None, :]
+    )
     rand = _curand_uniform_from_element_offset(
       philox_seed, philox_offset + linear
     )
